ScrapingGoogle/requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.44.3
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
soupsieve==2.5
urllib3==2.1.0
wordcloud==1.9.2
